Split Configuration extension into it's own independent extension package
