+++ .github/workflows/configuration-ci.yml
@@ -1,0 +1,26 @@
+name: CI
+
+on:
+  push:
+    branches: ['**']
+    paths: 
+      - Packages/Configuration/**
+      - .github/workflows/configuration-*
+
+jobs:
+  ci:
+    name: CI/CD
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Setup .Net
+        uses: actions/setup-dotnet@v3
+        with:
+          dotnet-version: |
+            7.0.x
+      - name: Build
+        working-directory: Packages/Configuration
+        run: dotnet build --configuration Release
+      - name: Test
+        working-directory: Packages/Configuration
+        run: dotnet test --configuration Release --no-build
