+++ .github/workflows/configuration-release.yml
@@ -1,0 +1,19 @@
+name: Release Configuration.Extension
+on:
+  pull_request:
+    types: [closed]
+    paths: 
+      - Packages/Configuration/**
+      - .github/workflows/configuration-*
+
+jobs:
+  release:
+    uses: ./.github/workflows/release.yml
+    with:
+      working-directory: Packages/Configuration
+      library-name: Configuration.Extension
+      dotnet-version: |
+        7.0.x
+      prerelease-branches: |
+        alpha
+    secrets: inherit
