+++ Packages/Configuration/Configuration.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Configuration.Extension", "Source\Configuration.Extension.csproj", "{FC9564BE-164C-4894-9B41-DFEF1B98B6F3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Configuration.Extension.Tests", "Tests\Configuration.Extension.Tests.csproj", "{F093DFB3-2184-439F-9CDA-8D34987A0D65}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{FC9564BE-164C-4894-9B41-DFEF1B98B6F3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FC9564BE-164C-4894-9B41-DFEF1B98B6F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FC9564BE-164C-4894-9B41-DFEF1B98B6F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FC9564BE-164C-4894-9B41-DFEF1B98B6F3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F093DFB3-2184-439F-9CDA-8D34987A0D65}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F093DFB3-2184-439F-9CDA-8D34987A0D65}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F093DFB3-2184-439F-9CDA-8D34987A0D65}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F093DFB3-2184-439F-9CDA-8D34987A0D65}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
