+++ Packages/Configuration/Directory.Build.props
@@ -1,6 +1,6 @@
 <Project>
 	<Import Project="$([MSBuild]::GetPathOfFileAbove('Directory.Build.props', '$(MSBuildThisFileDirectory)../'))" />
 	<PropertyGroup>
-        <PackageTags>$(PackageTags);inversionofcontrol;ioc;dependencyinjection;microsoft;autofac;lamar</PackageTags>
+        <PackageTags>$(PackageTags);inversionofcontrol;ioc;dependencyinjection;configuration;microsoft</PackageTags>
 	</PropertyGroup>
 </Project>
