+++ Packages/Configuration/README.md
@@ -1,0 +1,1 @@
+# Configuration.Extension
