+++ Packages/Configuration/Source/ConfigurationAttribute.cs
@@ -1,0 +1,31 @@
+// Copyright (c) woksin-org. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Configuration.Extension;
+
+/// <summary>
+/// Indicates that the type should be registered as a configuration object.
+/// </summary>
+[AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct)]
+// ReSharper disable once ClassNeverInstantiated.Global
+public class ConfigurationAttribute : Attribute
+{
+    /// <summary>
+    /// Initializes a new instance of the <see cref="ConfigurationAttribute"/> class.
+    /// </summary>
+    /// <param name="configurationPathParts">The configuration path parts to parse the object from, excluding the prefix that's configured with the <see cref="Settings"/> configuration.</param>
+    public ConfigurationAttribute(params string[] configurationPathParts)
+    {
+	    if (configurationPathParts.Length == 0)
+        {
+            throw new ArgumentException("Configuration attribute must include one or more sections", nameof(configurationPathParts));
+        }
+
+        ConfigurationPath = Microsoft.Extensions.Configuration.ConfigurationPath.Combine(configurationPathParts);
+    }
+    
+    /// <summary>
+    /// Gets the configuration path to parse the configuration object from.
+    /// </summary>
+    public string ConfigurationPath { get; }
+}
