+++ Packages/Configuration/Tests/Configuration.Extension.Tests.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <RootNamespace>Configuration.Extension</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Hosting" VersionOverride="7.0.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\Source\Configuration.Extension.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <None Include="appsettings_without_prefix.json" CopyToOutputDirectory="Always" />
+        <None Include="appsettings_with_prefix.json">
+          <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        </None>
+    </ItemGroup>
+
+</Project>
