+++ Packages/Configuration/version.json
@@ -1,0 +1,3 @@
+{
+    "version": "1.0.0-alpha.0"
+}
