--- Packages/IoC/Configuration.Tests/IoCExtensions.Configuration.Tests.csproj
@@ -1,24 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <RootNamespace>IoCExtensions.Configuration</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Hosting" VersionOverride="7.0.1" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\Configuration\IoCExtensions.Configuration.csproj" />
-        <ProjectReference Include="..\Autofac\IoCExtensions.Autofac.csproj" />
-        <ProjectReference Include="..\Lamar\IoCExtensions.Lamar.csproj" />
-        <ProjectReference Include="..\Microsoft\IoCExtensions.Microsoft.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <None Include="appsettings_without_prefix.json" CopyToOutputDirectory="Always" />
-        <None Include="appsettings_with_prefix.json">
-          <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-        </None>
-    </ItemGroup>
-
-</Project>
