--- Packages/IoC/Configuration/IoCExtensions.Configuration.csproj
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Configuration" />
-        <PackageReference Include="Microsoft.Extensions.Configuration.Json" />
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
-        <PackageReference Include="Microsoft.Extensions.FileProviders.Abstractions" />
-        <PackageReference Include="Microsoft.Extensions.FileProviders.Physical" />
-        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
-        <PackageReference Include="Microsoft.Extensions.Options" />
-        <PackageReference Include="NetEscapades.Configuration.Yaml" />
-        <PackageReference Include="Newtonsoft.Json" />
-    </ItemGroup>
-    
-	<ItemGroup>
-		<ProjectReference Include="..\Core\IoCExtensions.Core.csproj" />
-	</ItemGroup>
-    
-    <ItemGroup>
-        <None Include="$(MSBuildThisFileDirectory)/README.md" Pack="true" PackagePath="/README.md" />
-    </ItemGroup>
-</Project>
