--- Packages/IoC/Configuration/README.md
@@ -1,1 +1,0 @@
-# IoCExtensions.Configuration
