+++ Packages/IoC/IoC.sln
@@ -5,10 +5,6 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IoCExtensions.Autofac", "Autofac\IoCExtensions.Autofac.csproj", "{D822BF13-0D45-4EC3-B3BB-879FA04A241A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IoCExtensions.Configuration.Tests", "Configuration.Tests\IoCExtensions.Configuration.Tests.csproj", "{D488EFEA-9D00-44D5-9952-30F16007D2D1}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IoCExtensions.Configuration", "Configuration\IoCExtensions.Configuration.csproj", "{26E560B5-E0C8-4E50-BEC8-7366FAA6CC24}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IoCExtensions.Core", "Core\IoCExtensions.Core.csproj", "{E53EBC5D-72E5-4B2C-B1F2-00D3C3F6B8CF}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IoCExtensions.Lamar", "Lamar\IoCExtensions.Lamar.csproj", "{5EBECF37-940D-41C4-B56C-3353B525C7BB}"
@@ -30,14 +26,6 @@ Global
 		{D822BF13-0D45-4EC3-B3BB-879FA04A241A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D822BF13-0D45-4EC3-B3BB-879FA04A241A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D822BF13-0D45-4EC3-B3BB-879FA04A241A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D488EFEA-9D00-44D5-9952-30F16007D2D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D488EFEA-9D00-44D5-9952-30F16007D2D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D488EFEA-9D00-44D5-9952-30F16007D2D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D488EFEA-9D00-44D5-9952-30F16007D2D1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{26E560B5-E0C8-4E50-BEC8-7366FAA6CC24}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{26E560B5-E0C8-4E50-BEC8-7366FAA6CC24}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{26E560B5-E0C8-4E50-BEC8-7366FAA6CC24}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{26E560B5-E0C8-4E50-BEC8-7366FAA6CC24}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E53EBC5D-72E5-4B2C-B1F2-00D3C3F6B8CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E53EBC5D-72E5-4B2C-B1F2-00D3C3F6B8CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E53EBC5D-72E5-4B2C-B1F2-00D3C3F6B8CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
